Create ChartControls component
Moves the chart controls into a component that wraps the selected charts for reusability. If it's unintuitive to use I may add a comment to explain it.
